requirements.txt
feedparser==6.0.8
Flask==2.0.2
flask_cors==3.0.10
python-dotenv==0.19.2
transformers==4.16.2
torch